Added layout section for Bootstrap modals.
Bootstrap modals must be defined as a direct child of the body tag. They cannot be placed in the standard 'content' section because every element of that section is the children of several divs with varying styles important to the master layout. This breaks the Bootstrap modal. The new section is located just before the closing body tag in the master layout.